[Docs]: 필요없는 코드 주석처리 및 삭제
