+++ Assets/Programing/LYJ/Scripts/InGameItemReference.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-//아이템에 이 스크립트를 붙ㅇ 사용
+//아이템에 이 스크립트를 붙여 사용
 //해당 아이템이 가지고 있는 데이터를 쉽게 참조할 수 있음
 public class InGameItemReference : MonoBehaviour
 {
