+++ Assets/Programing/LYJ/TestPlayer.cs
@@ -9,8 +9,6 @@ public class TestPlayer : MonoBehaviour
 
         if (itemReference != null)
         {
-            // 아이템을 수집
-            StatusWindowController.Instance.CollectItem(itemReference.item);
 
             Destroy(other.gameObject);
